Replaced old test with new test using mocked MemoryManager
